services/processing-requests/pyproject.toml
[tool.poetry]
name = "processing-requests"
version = "0.1.0"
description = ""
authors = ["Ihar Suvorau <ihar.suvorau@gmail.com>"]
readme = "README.md"
packages = [{ include = "processing_requests" }]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.20"
fastapi = "^0.103.1"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
python-multipart = "^0.0.6"
greenlet = "^2.0.2"
opentelemetry-distro = "^0.41b0"
opentelemetry-exporter-otlp = "^1.20.0"
opentelemetry-instrumentation-fastapi = "^0.41b0"
opentelemetry-instrumentation-httpx = "^0.41b0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
